I added the ability to pass in a label for the 'Existing Tags'.
I scoped the click event on the tag to be specifically on the anchor. After I changed the tag markup somewhat, the click event was on the whole tag.

Added a class name to the tag container div so the popular tags container can be differentiated from the existing tags container. 



git-svn-id: https://svn.symfony-project.com/plugins/sfDoctrineActAsTaggablePlugin/trunk@30820 ee427ae8-e902-0410-961c-c3ed070cd9f9